[ty] Move venv and conda env discovery to `SearchPath::from_settings`
