+++ crates/ruff_benchmark/benches/ty_walltime.rs
@@ -6,10 +6,11 @@ use rayon::ThreadPoolBuilder;
 use ruff_benchmark::real_world_projects::{InstalledProject, RealWorldProject};
 use ruff_db::system::{OsSystem, SystemPath, SystemPathBuf};
 
+use ruff_db::ranged_value::RangedValue;
 use ruff_db::testing::setup_logging_with_filter;
 use ruff_python_ast::PythonVersion;
 use ty_project::metadata::options::{EnvironmentOptions, Options};
-use ty_project::metadata::value::{RangedValue, RelativePathBuf};
+use ty_project::metadata::value::RelativePathBuf;
 use ty_project::{Db, ProjectDatabase, ProjectMetadata};
 
 struct Benchmark<'a> {
