+++ crates/ruff_db/src/lib.rs
@@ -17,6 +17,7 @@ pub mod system;
 #[cfg(feature = "testing")]
 pub mod testing;
 pub mod vendored;
+pub mod ranged_value;
 
 #[cfg(not(target_arch = "wasm32"))]
 pub use std::time::{Instant, SystemTime, SystemTimeError};
