+++ crates/ruff_graph/src/db.rs
@@ -3,6 +3,7 @@ use std::sync::Arc;
 use zip::CompressionMethod;
 
 use ruff_db::files::{File, Files};
+use ruff_db::ranged_value::RangedValue;
 use ruff_db::system::{OsSystem, System, SystemPathBuf};
 use ruff_db::vendored::{VendoredFileSystem, VendoredFileSystemBuilder};
 use ruff_db::{Db as SourceDb, Upcast};
@@ -10,7 +11,7 @@ use ruff_python_ast::PythonVersion;
 use ty_python_semantic::lint::{LintRegistry, RuleSelection};
 use ty_python_semantic::{
     Db, Program, ProgramSettings, PythonPath, PythonPlatform, PythonVersionSource,
-    PythonVersionWithSource, SearchPathSettings, SysPrefixPathOrigin, default_lint_registry,
+    PythonVersionWithSource, SearchPathSettings, default_lint_registry,
 };
 
 static EMPTY_VENDORED: std::sync::LazyLock<VendoredFileSystem> = std::sync::LazyLock::new(|| {
@@ -36,11 +37,10 @@ impl ModuleDb {
         venv_path: Option<SystemPathBuf>,
     ) -> Result<Self> {
         let mut search_paths = SearchPathSettings::new(src_roots);
+        // TODO: Consider setting `PythonPath::Auto` if no venv_path is provided.
         if let Some(venv_path) = venv_path {
-            search_paths.python_path =
-                PythonPath::sys_prefix(venv_path, SysPrefixPathOrigin::PythonCliFlag);
-        }
-
+            search_paths.python_path = PythonPath::IntoSysPrefix(RangedValue::cli(venv_path));
+        };
         let db = Self::default();
         let search_paths = search_paths
             .to_search_paths(db.system(), db.vendored())
