+++ crates/ty_project/src/metadata/options.rs
@@ -2,10 +2,10 @@ use crate::Db;
 use crate::combine::Combine;
 use crate::glob::{ExcludeFilter, IncludeExcludeFilter, IncludeFilter, PortableGlobKind};
 use crate::metadata::settings::{OverrideSettings, SrcSettings};
-use crate::metadata::value::{
-    RangedValue, RelativeGlobPattern, RelativePathBuf, ValueSource, ValueSourceGuard,
-};
+use ruff_db::ranged_value::{RangedValue, ValueSource, ValueSourceGuard};
 
+use super::settings::{Override, Settings, TerminalSettings};
+use crate::metadata::value::{RelativeGlobPattern, RelativePathBuf};
 use ordermap::OrderMap;
 use ruff_db::RustDoc;
 use ruff_db::diagnostic::{
@@ -30,10 +30,7 @@ use ty_python_semantic::lint::{GetLintError, Level, LintSource, RuleSelection};
 use ty_python_semantic::{
     ProgramSettings, PythonPath, PythonPlatform, PythonVersionFileSource, PythonVersionSource,
     PythonVersionWithSource, SearchPathSettings, SearchPathValidationError, SearchPaths,
-    SysPrefixPathOrigin,
 };
-
-use super::settings::{Override, Settings, TerminalSettings};
 
 #[derive(
     Debug, Default, Clone, PartialEq, Eq, Combine, Serialize, Deserialize, OptionsMetadata,
@@ -235,31 +232,9 @@ impl Options {
                 .python
                 .as_ref()
                 .map(|python_path| {
-                    let origin = match python_path.source() {
-                        ValueSource::Cli => SysPrefixPathOrigin::PythonCliFlag,
-                        ValueSource::File(path) => SysPrefixPathOrigin::ConfigFileSetting(
-                            path.clone(),
-                            python_path.range(),
-                        ),
-                    };
-                    PythonPath::sys_prefix(python_path.absolute(project_root, system), origin)
+                    PythonPath::IntoSysPrefix(python_path.absolute_ranged(project_root, system))
                 })
-                .or_else(|| {
-                    system.env_var("VIRTUAL_ENV").ok().map(|virtual_env| {
-                        PythonPath::sys_prefix(virtual_env, SysPrefixPathOrigin::VirtualEnvVar)
-                    })
-                })
-                .or_else(|| {
-                    system.env_var("CONDA_PREFIX").ok().map(|path| {
-                        PythonPath::sys_prefix(path, SysPrefixPathOrigin::CondaPrefixVar)
-                    })
-                })
-                .unwrap_or_else(|| {
-                    PythonPath::sys_prefix(
-                        project_root.to_path_buf(),
-                        SysPrefixPathOrigin::LocalVenv,
-                    )
-                }),
+                .unwrap_or_else(|| PythonPath::Auto(project_root.to_path_buf())),
         };
 
         settings.to_search_paths(system, vendored)
@@ -1033,7 +1008,7 @@ pub struct TerminalOptions {
 #[derive(Debug, Default, Clone, PartialEq, Eq, Combine, Serialize, Deserialize, RustDoc)]
 #[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
 #[serde(transparent)]
-pub struct OverridesOptions(Vec<RangedValue<OverrideOptions>>);
+pub struct OverridesOptions(Vec<RangedOverrideOptions>);
 
 impl OptionsMetadata for OverridesOptions {
     fn documentation() -> Option<&'static str> {
@@ -1046,7 +1021,7 @@ impl OptionsMetadata for OverridesOptions {
 }
 
 impl Deref for OverridesOptions {
-    type Target = [RangedValue<OverrideOptions>];
+    type Target = [RangedOverrideOptions];
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -1122,7 +1097,12 @@ pub struct OverrideOptions {
     pub rules: Option<Rules>,
 }
 
-impl RangedValue<OverrideOptions> {
+#[derive(serde::Deserialize, Debug, Clone, PartialEq, Eq, Combine, serde::Serialize)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
+#[serde(transparent)]
+pub struct RangedOverrideOptions(RangedValue<OverrideOptions>);
+
+impl RangedOverrideOptions {
     fn to_override(
         &self,
         db: &dyn Db,
@@ -1130,7 +1110,7 @@ impl RangedValue<OverrideOptions> {
         global_rules: Option<&Rules>,
         diagnostics: &mut Vec<OptionDiagnostic>,
     ) -> Result<Option<Override>, Box<OptionDiagnostic>> {
-        let rules = self.rules.or_default();
+        let rules = self.0.rules.or_default();
 
         // First, warn about incorrect or useless overrides.
         if rules.is_empty() {
@@ -1140,7 +1120,7 @@ impl RangedValue<OverrideOptions> {
                 Severity::Warning,
             );
 
-            diagnostic = if self.rules.is_none() {
+            diagnostic = if self.0.rules.is_none() {
                 diagnostic = diagnostic.sub(SubDiagnostic::new(
                     Severity::Info,
                     "It has no `rules` table",
@@ -1166,10 +1146,10 @@ impl RangedValue<OverrideOptions> {
             ));
 
             // Add source annotation if we have source information
-            if let Some(source_file) = self.source().file() {
+            if let Some(source_file) = self.0.source().file() {
                 if let Ok(file) = system_path_to_file(db.upcast(), source_file) {
                     let annotation =
-                        Annotation::primary(Span::from(file).with_optional_range(self.range()))
+                        Annotation::primary(Span::from(file).with_optional_range(self.0.range()))
                             .message("This overrides section configures no rules");
                     diagnostic = diagnostic.with_annotation(Some(annotation));
                 }
@@ -1180,8 +1160,9 @@ impl RangedValue<OverrideOptions> {
             return Ok(None);
         }
 
-        let include_missing = self.include.is_none();
+        let include_missing = self.0.include.is_none();
         let exclude_empty = self
+            .0
             .exclude
             .as_ref()
             .is_none_or(|exclude| exclude.is_empty());
@@ -1194,7 +1175,7 @@ impl RangedValue<OverrideOptions> {
                 Severity::Warning,
             );
 
-            diagnostic = if self.exclude.is_none() {
+            diagnostic = if self.0.exclude.is_none() {
                 diagnostic.sub(SubDiagnostic::new(
                     Severity::Info,
                     "It has no `include` or `exclude` option restricting the files",
@@ -1217,10 +1198,10 @@ impl RangedValue<OverrideOptions> {
             ));
 
             // Add source annotation if we have source information
-            if let Some(source_file) = self.source().file() {
+            if let Some(source_file) = self.0.source().file() {
                 if let Ok(file) = system_path_to_file(db.upcast(), source_file) {
                     let annotation =
-                        Annotation::primary(Span::from(file).with_optional_range(self.range()))
+                        Annotation::primary(Span::from(file).with_optional_range(self.0.range()))
                             .message("This overrides section applies to all files");
                     diagnostic = diagnostic.with_annotation(Some(annotation));
                 }
@@ -1234,7 +1215,7 @@ impl RangedValue<OverrideOptions> {
         let include = build_include_filter(
             db,
             project_root,
-            self.include.as_ref(),
+            self.0.include.as_ref(),
             GlobFilterContext::Overrides,
             diagnostics,
         )?;
@@ -1242,7 +1223,7 @@ impl RangedValue<OverrideOptions> {
         let exclude = build_exclude_filter(
             db,
             project_root,
-            self.exclude.as_ref(),
+            self.0.exclude.as_ref(),
             &[],
             GlobFilterContext::Overrides,
         )?;
@@ -1262,7 +1243,7 @@ impl RangedValue<OverrideOptions> {
         let override_instance = Override {
             files,
             options: Arc::new(InnerOverrideOptions {
-                rules: self.rules.clone(),
+                rules: self.0.rules.clone(),
             }),
             settings: Arc::new(OverrideSettings {
                 rules: rule_selection,
